==> notebooks/1.0_test_repartition_categories_train_test_val_sets.ipynb <==
import pandas as pd
import numpy as np
import plotly.graph_objects as go
import os
import sys
from pathlib import Path

project_root = Path().resolve().parent
if not project_root in [Path(p).resolve() for p in sys.path]:
    sys.path.append(str(project_root))
from src import PATHS

df_train  = pd.read_csv(os.path.join(PATHS.raw_data, 'labels', 'train.txt'), sep=" ", header=None)
df_train.columns=['path', 'label']
df_train['set']=['train' for i in range (len(df_train))]

df_test  = pd.read_csv(os.path.join(PATHS.raw_data, 'labels', 'test.txt'), sep=" ", header=None)
df_test.columns=['path', 'label']
df_test['set']=['test' for i in range (len(df_test))]

df_val  = pd.read_csv(os.path.join(PATHS.raw_data, 'labels', 'val.txt'), sep=" ", header=None)
df_val.columns=['path', 'label']
df_val['set']=['val' for i in range (len(df_val))]

#df_sets = pd.concat([df_train, df_test, df_val],ignore_index=True)
#df_sets.to_parquet(os.path.join(PATHS.metadata, "df_path_label_set.parquet"), index=False)

compte=pd.concat([pd.DataFrame(df_train['cat'].value_counts()).sort_index(),
                  pd.DataFrame(df_test['cat'].value_counts()).sort_index(),
                  pd.DataFrame(df_val['cat'].value_counts()).sort_index()], axis=1)
compte.columns=['train', 'test', 'val']
display(compte)



fig = go.Figure(data=[
    go.Bar(name='Train', x=compte.index, y=compte['train'], marker_color='rgb(87, 24, 69)'),
    go.Bar(name='Test', x=compte.index, y=compte['test'], marker_color='rgb(199, 0, 57)'),
    go.Bar(name='Validation', x=compte.index, y=compte['val'], marker_color='rgb(255,87,51)'),
])
# Change the bar mode
fig.update_layout(
    barmode='stack', 
    title=dict(text="Répartition des données dans les 3 subsets"),
    xaxis=dict(title=dict(text="Catégories"), 
               tickmode='linear',  # Mode linéaire
               dtick=1             # Un tick toutes les 1 unité
              ),
    yaxis=dict(title=dict(text="Nombre de documents")),
    legend=dict(title=dict(text="Subset"))
)
fig.show()
