==> notebooks/1.5-image_completion.ipynb <==
## 1. Préparation

import sys
from pathlib import Path

project_root = Path().resolve().parent
if not project_root in [Path(p).resolve() for p in sys.path]:
    sys.path.append(str(project_root))

from src import PATHS
from src.data_completion import get_images_to_fix, convert_iit_to_rvl_tiff

import pandas as pd

## 2. Complétion

### Vérification des images

df = pd.read_parquet(PATHS.metadata / "df_filepaths.parquet")[["rvl_image_path", "iit_image_path"]]
df["rvl"] = PATHS.data / df.rvl_image_path
df["iit"] = PATHS.data / df.iit_image_path
image_paths = df[["rvl", "iit"]]
image_paths.head(2)

to_repair = get_images_to_fix(df.rvl)
to_repair

### Recréation des images manquantes à partir d'IIT CDIP

for rvl_file_path in to_repair:
    iit_file_path = image_paths[image_paths.rvl == rvl_file_path].iit.iloc[0]
    convert_iit_to_rvl_tiff(iit_file_path, rvl_file_path)
