==> notebooks/2.4_df-texte-et-completion-ocr.ipynb <==
# Création DataFrame IIT_CDIP_COLL_XML_FEATURES

## README
Après comparaison des différentes dataframes, il apparait que la base de données ayant la meilleure complétude et les odnnées naturellement les plus propres est celle contenue dans df_iit_cdip_coll_xml_ltdlwocr_features.

En choisissant de ne traiter un document que sur les seules informations qu'il contient (et donc sans les métadonnées telles que son emetteur, son récipiendaire, ...), nous avons choisi de ne concerver que l'ocr.

Ce notebook permet de créer un dataframe qui contiendra ces seules données (chapitre 2).

Dans un deuxième temps (chapitre 3), les données manquantes seront complétées par utilisation d'un outil d'océrisation sur les images IIT-CDIP (meilleure résolution que RVL-CDIP) correspondantes.


## 1. Préparation

import sys
from pathlib import Path

project_root = Path().resolve().parent
if not project_root in [Path(p).resolve() for p in sys.path]:
    sys.path.append(str(project_root))

from src import PATHS

import pandas as pd
from PIL import Image
import pytesseract
from tqdm import tqdm

## 2. Préparation de la base de données

df_ocr = pd.read_parquet(PATHS.processed_data / "df_iit_cdip_coll_xml_ltdlwocr_features.parquet")[["ocr_output"]]
df_ocr.isna().sum()

## 3. Complétion des ocr manquants

df_documents = pd.read_parquet(PATHS.metadata / "df_documents.parquet")
missing_ocrs = df_documents[df_documents.join(df_ocr).ocr_output.isna()]

image_paths = PATHS.data / pd.read_parquet(PATHS.metadata / "df_filepaths.parquet").iit_image_path

import pytesseract
try:
    version = pytesseract.get_tesseract_version()
    print("Tesseract version:", version)
    print("pytesseract est bien configuré.")
except Exception as e:
    print("Erreur:", e)

# EN CAS D'ERREUR
# Si tesseract est bien installé mais ne se trouve pas dans le path, il est possible de définir son chemin avec cette commande:
pytesseract.pytesseract.tesseract_cmd = "/opt/homebrew/bin/tesseract" # Chemin à mettre à jour le cas echeant

try:
    version = pytesseract.get_tesseract_version()
    print("Tesseract version:", version)
    print("pytesseract est bien configuré.")
except Exception as e:
    print("Erreur:", e)

def extract_text_from_image(image_path: str) -> str:
    image = Image.open(image_path)
    text = pytesseract.image_to_string(image)
    return text

# Environ 3h00 de calcul pour 13346 lignes
ocrs = []
for file_path in tqdm(image_paths[missing_ocrs.index].values):
    ocrs.append(extract_text_from_image(file_path))
pd.DataFrame(ocrs, index=missing_ocrs.index)

recomputed_ocrs = pd.DataFrame(ocrs, index=missing_ocrs.index, columns = ["ocr_output"])

df_ocr = df_ocr.combine_first(recomputed_ocrs).rename(columns={"ocr_output":"raw_ocr"})
df_ocr.isna().sum()

len(df_ocr)

### 3.2. Sauvegarde des DataFrames

df_ocr.to_parquet(PATHS.processed_data / "df_raw_ocr.parquet")
