==> notebooks/3.6-visualisation-de-texte.ipynb <==
# Dataviz texte
Ce notebook crée un nuage de mots sur les textes océrisés

## 1. Préparation

import sys
from pathlib import Path

project_root = Path().resolve().parent
if not project_root in [Path(p).resolve() for p in sys.path]:
    sys.path.append(str(project_root))

from src import PATHS

import pandas as pd
from matplotlib import pyplot as plt
from wordcloud import WordCloud

## 2. Sur les données brutes

serie = pd.read_parquet(PATHS.processed_data / "df_raw_ocr.parquet").raw_ocr.dropna()
serie.head()

text = " ".join(serie.sample(10000).values)

wordcloud = WordCloud(width=800, height=400, background_color='white').generate(text)

plt.figure(figsize=(20, 10))
plt.imshow(wordcloud, interpolation='bilinear')
plt.axis('off')
plt.title("Nuage de mots sur les données ocr brutes")
plt.show()

# on observe la prédominance de mots de 1 lettres et de l'expression "pgNbr", 
# qui est un artéfact de numérisation ou d'océrisation


## 2. Sur les données préprocessées
Pour exécuter cette partie, il faut avoir exécuté le notebook 4.3 (text preprocessing)

serie = pd.read_parquet(PATHS.processed_data / "df_txt_ocr1.parquet").ocr.dropna()
serie.head()

text = " ".join(serie.sample(10000).values)

wordcloud = WordCloud(width=800, height=400, background_color='white').generate(text)

plt.figure(figsize=(20, 10))
plt.imshow(wordcloud, interpolation='bilinear')
plt.axis('off')
plt.title("Nuage de mots sur les données ocr brutes")
plt.show()

# il reste des mots de 2 lettres qui ne semblent pas avoir vraiment de sens
# mais on observe beaucoup moins de pollution par des artéfacts d'océrisation
