==> notebooks/8.1 Test des model_wrappers.ipynb <==
import sys
from pathlib import Path
import pandas as pd
from tqdm import tqdm
project_root = Path().resolve().parent
if not project_root in [Path(p).resolve() for p in sys.path]:
    sys.path.append(str(project_root))

from src import PATHS

from src.models.model_wrappers import ModelWrapperFactory

documents = pd.read_parquet(PATHS.metadata / "df_documents.parquet")
labels = pd.read_parquet(PATHS.metadata / "df_encoded_labels.parquet").label

### _________________________________________________________________________________________________________________________________________________________________________________________________________________

# TEST

registered = ModelWrapperFactory.get_registered()
registered

# Tableau de predictions

# to_load = registered
to_load = [
    'Image-based ResNet50 frozen',
    'Image-based VGG16 unfreezed',
    'Image-based LGBM',
    'Text-based Logistic Regressor',
    'Text-based MLP1',
    'Text-based MLP3',
    'MMO-Composite LogReg on img-LGBM+txt-LogReg',
    'MMO-CLIP-Based Logistic Regressor',
    'MMO-CLIP-Based MLP12'
]
model_wrappers = [ModelWrapperFactory.load_existing(mw) for mw in tqdm(to_load)]

# Predictions sur un extrait de RVL-CDIP
document_sample = documents.sample(50).index
results = [f"{m.name:45}" + "|".join([f"{i:2d}" for i in m.predict(document_sample)]) for m in tqdm(model_wrappers)]

# Affichage tableau de synthèse
print(f"{'Label':45}" + "|".join([f"{i:2d}" for i in labels[document_sample].values]))
print("-"*194)    
for r in results:
    print(r)

# Predictions sur des documents importés (il faut avoir utilisé le notebook 8.2 avant
import pandas as pd
import numpy as np

documents = pd.read_parquet(PATHS.metadata / "df_documents.parquet").index
documents_rvl = pd.read_parquet(PATHS.metadata / "df_documents_save.parquet").index
imported_ids = [i for i in documents if i not in documents_rvl]


if len(imported_ids) > 50:
    document_sample = np.random.choice(imported_ids, 50)
else:
    document_sample = imported_ids
if imported_ids:
    results = [f"{m.name:45}" + "|".join([f"{i:2d}" for i in m.predict(document_sample)]) for m in tqdm(model_wrappers)]
    
    # Affichage tableau de synthèse
    for r in results:
        print(r)
else:
    print("Aucun document à prédire. Si vous avez fait un import récent, pensez à relancer le noyau de ce notebook pour que la liste des documents importés soit mise à jour.")

### _________________________________________________________________________________________________________________________________________________________________________________________________________________

# Calcul de tous les performance summary
Pour une premiere inférence, peut se révéler très long (40K inférence par modele), notamment sur les modeles img-DL (jusqu'à près de 2h par modèle)
une fois calculés, les performance summary sont automatiquement sauveagardés dans le dossier models

for name in tqdm(registered):
    ModelWrapperFactory.load_existing(name).performance_summary
