==> notebooks/8.2-Prédiction-image-externe.ipynb <==
import sys
from pathlib import Path

project_root = Path().resolve().parent
if not project_root in [Path(p).resolve() for p in sys.path]:
    sys.path.append(str(project_root))

from src import PATHS, LABELS

from src.models.preprocessing import extract_features, reset_to_rvl_state

# Import

for i in PATHS.images_to_import.iterdir():
    extract_features(i)

# Nettoyage

reset_to_rvl_state()

# Réduction DF (pour démo)
etant donnée la méthode utilisée pour faire la démo (réutilisation des DF RVL), les temps d'écritures sur les DF seraient trop long.
On réduit donc le jeu de données à 4000 lignes en appliquant une fois ce script (penser a faire une copie manuelle de df_documents.parquet en df_documents_save.parquet pour pouvoir utiliser reset_to_rvl_state ensuite)

!!! Attention: bien sur faire une copie séparée de processed et metadata avant

import shutil
import pandas as pd


# code de reset_rvl_state

rvl_ids = pd.read_parquet(PATHS.samples / "df_documents_sample_4k_1.parquet").index
for df_path in [
    PATHS.metadata / "df_documents.parquet",
    PATHS.processed_data / "df_raw_ocr.parquet",
    PATHS.processed_data / "df_txt_ocr1.parquet",
    PATHS.processed_data / "df_img_features_pixels.parquet",
    PATHS.processed_data / "df_clip_embeddings.parquet",
    
    PATHS.processed_data / "df_clip_embeddings.parquet",
    PATHS.metadata / "df_data_sets.parquet",
    PATHS.metadata / "df_plain_labels.parquet",
    PATHS.metadata / "df_encoded_labels.parquet",
    PATHS.metadata / "df_filepaths.parquet",
    
    
    
    
]:
    print(df_path, end="...")
    df = pd.read_parquet(df_path)
    df = df.loc[rvl_ids,:]
    assert len(df) == 4000
    df.to_parquet(df_path)
    print(" sucessfully reset")
    rvl_path = PATHS.other_images
    converted_path = Path(str(rvl_path).replace(str(PATHS.rvl_cdip_images), str(PATHS.converted_images)))
for folder in [str(rvl_path), str(converted_path)]:
    pass
    try:
        print(str(folder), end="...")
        shutil.rmtree(folder)
        print(" sucessfully removed")
    except Exception as e:
        print(f"⚠️ Erreur lors de la suppression de {folder}: {e}")
